feat: make dialog import a menu to select more options to import
