+++ src/lib/layouts/rest/views/index.ts
@@ -1,5 +1,6 @@
 export * from './view-welcome';
 export * from './view-instructions';
+export * from './view-import-from-file';
 export * from './view-collections';
 export * from './view-select-collections';
 export * from './view-edit';
