+++ src/lib/layouts/rest/views/view-import-from-file/index.ts
@@ -1,0 +1,3 @@
+import { default as ViewImportFromFile } from './view-import-from-file.layout.svelte';
+export { ViewImportFromFile };
+export default ViewImportFromFile;
